refactor: utils functions seprated by function types
