+++ hashes/index.js
@@ -1,0 +1,7 @@
+const mimcHash = require('./mimc');
+const shaHash = require('./sha256');
+
+module.exports = {
+  shaHash,
+  mimcHash,
+};
