+++ hashes/mimc/config.js
@@ -1,0 +1,21 @@
+module.exports = {
+  BLS12_377: {
+    exponent: 11,
+    rounds: 74,
+    modulus: BigInt('8444461749428370424248824938781546531375899335154063827935233455917409239041'),
+  },
+  ALT_BN128: {
+    exponent: 7,
+    rounds: 91,
+    modulus: BigInt(
+      '21888242871839275222246405745257275088548364400416034343698204186575808495617',
+    ),
+  },
+  BW6_761: {
+    exponent: 23,
+    rounds: 84,
+    modulus: BigInt(
+      '258664426012969094010652733694893533536393512754914660539884262666720468348340822774968888139573360124440321458177',
+    ),
+  },
+};
